Move VirtualClient to interfaces
