+++ python/pyrogue/gui/__main__.py
@@ -10,6 +10,7 @@
 
 import argparse
 import pyrogue
+import pyrogue.interfaces
 import pyrogue.gui
 import sys
 
@@ -22,7 +23,7 @@ args = parser.parse_args()
 print("Connecting to host {} port {}".format(args.host,args.port))
 
 # Connect to the server
-client = pyrogue.VirtualClient(args.host,args.port)
+client = pyrogue.interfaces.VirtualClient(args.host,args.port)
 
 # Create GUI
 appTop = pyrogue.gui.application(sys.argv)
