+++ python/pyrogue/gui/_gui.py
@@ -27,6 +27,7 @@ except ImportError:
     from PyQt4.QtGui     import *
 
 import pyrogue
+import pyrogue.interfaces
 import pyrogue.gui
 import pyrogue.gui.variables
 import pyrogue.gui.commands
@@ -41,7 +42,7 @@ def application(argv):
 class GuiTop(QWidget):
 
     newRoot = pyqtSignal(pyrogue.Root,int)
-    newVirt = pyqtSignal(pyrogue.VirtualNode,int)
+    newVirt = pyqtSignal(pyrogue.interfaces.VirtualNode,int)
 
     def __init__(self,*, parent=None, minVisibility=25, group=None):
         super(GuiTop,self).__init__(parent)
@@ -78,13 +79,13 @@ class GuiTop(QWidget):
         if not root.running:
             raise Exception("GUI can not be attached to a tree which is not started")
 
-        if isinstance(root,pyrogue.VirtualNode):
+        if isinstance(root,pyrogue.interfaces.VirtualNode):
             self.newVirt.emit(root,self._minVisibility)
         else:
             self.newRoot.emit(root,self._minVisibility)
 
     @pyqtSlot(pyrogue.Root,int)
-    @pyqtSlot(pyrogue.VirtualNode,int)
+    @pyqtSlot(pyrogue.interfaces.VirtualNode,int)
     def _addTree(self,root,minVisibility):
         self.sys = pyrogue.gui.system.SystemWidget(root=root,parent=self.tab)
         self.tab.addTab(self.sys,root.name)
