+++ python/pyrogue/interfaces/_Virtual.py
@@ -24,15 +24,6 @@ import functools as ft
 import jsonpickle
 
 
-def genBaseList(cls):
-    ret = [str(cls)]
-
-    for l in cls.__bases__:
-        if l is not object:
-            ret += genBaseList(l)
-
-    return ret
-
 
 class VirtualProperty(object):
     def __init__(self, node, attr):
