+++ python/pyrogue/interfaces/__init__.py
@@ -9,4 +9,5 @@
 #-----------------------------------------------------------------------------
 
 from pyrogue.interfaces._ZmqServer import *
+from pyrogue.interfaces._Virtual   import *
